move pthread related functions
